Update OneDeviceAcquisitionExample.py
Simplified and generalized plux.pyd selection: now it works for every OS and python version.
This same update could be done in MultipleDeviceThreadingExample.py.
(Only tested in a 64bit Windows with python 3.9 and 3.10. It's just a folder selection, it should work in the remaining cases... but you know you bugs work).

Renamed variable `time` for `duration`.

Simplified NewDevice.onRawFrame() and now it prints sequence number and all the data values.

Running this as a script (__name__ == '__main__') will run the example, importing it wont.

Now the arguments from the command line are used as arguments for exampleAcquisition(). The following command:
~$ python OneDeviceAcquisitionExample.py   BTH00:07:80:4D:2E:76   60   200
will call exampleAcquisition("BTH00:07:80:4D:2E:76", "60", "200", code=0x01)  # code uses the default value because it hasn't been specified in the command line.
This means that exampleAcquisition() has to make sure to convert the arguments from str to int.